+++ src/core/default/backend.py
@@ -1,25 +1,22 @@
 import json
 import os
 
-from pathlib import PosixPath, WindowsPath
 from pydantic.main import BaseModel
 from pydantic.types import DirectoryPath, FilePath
 
 from typing import Dict, List, Any, Tuple
 import uuid
 
-from core.constructs.workspace import Workspace
 
+from core.constructs.backend import Backend_Configuration, Backend
+from core.constructs.backend_exceptions import *
 
-from ..constructs.backend import Backend_Configuration, Backend
-from ..constructs.backend_exceptions import *
-
-from ..constructs.components import (
+from core.constructs.components import (
     Component_Change_Type,
     ComponentModel,
     Component_Difference,
 )
-from ..constructs.resource import (
+from core.constructs.resource import (
     Resource_Change_Type,
     Resource_Difference,
     Resource_Reference_Change_Type,
@@ -27,7 +24,7 @@ from ..constructs.resource import (
     Resource_Reference_Difference,
     ResourceReferenceModel,
 )
-from ..constructs.resource_state import Resource_State
+from core.constructs.resource_state import Resource_State
 from ..utils import file_manager, hasher as cdev_hasher